Don't read AiProtection DB if zero bytes
If banaiprotect is enabled, but AiProtection has never been enabled
on a router, the sqlite DB file will exist but with zero bytes.

This causes error messages during an interactive banmalware refresh
since the tables do not exist in the empty DB file.

Only query the tables if the DB file has a non-zero file size.